+++ src/Soluble/Japha/Bridge/Exception/AuthenticationException.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Soluble Japha
+ *
+ * @link      https://github.com/belgattitude/soluble-japha
+ * @copyright Copyright (c) 2013-2017 Vanvelthem Sébastien
+ * @license   MIT License https://github.com/belgattitude/soluble-japha/blob/master/LICENSE.md
+ */
+
+namespace Soluble\Japha\Bridge\Exception;
+
+class AuthenticationException extends ConnectionException
+{
+}